Updates to the driving logic, behaviors are finished
The behaviors for the car are finished, now I just need to improve the driving logic of the car. To do that I'm following the TORCS official tutorial and adapting it to work with my behavior tree.

The car makes it around the track (sometimes), but tends to get on the grass/sand which causes it to spin out. Perhaps it wants to be a rally car.