Unstuck Node, additional improvements to the AI
While still following the TORCS tutorial, I have implemented some improvements to give the car more competitive times. I've tried implementing the filterTrack() method, but it causes the car to fly away (is this the future?).

Car makes it around the Forza track, with somewhat competitive time. It doesn't go on grass/sand anymore, so it actually stays on track.

Added a check to the Acceleration, Brake and Gear nodes, to make sure the car is not stuck; because when the AI detects the car is stuck, the StuckNode fully takes control of the car to get it back on track.

The Visual Studio Solution is organised with Filters/Folders (if they work).